refactor: refactor households
