+++ app/dashboard/households/shared/components/PaginationControls.tsx
@@ -1,0 +1,95 @@
+'use client'
+
+import { ChevronLeft, ChevronRight } from 'lucide-react'
+
+interface PaginationControlsProps {
+  currentPage: number
+  totalPages: number
+  itemsPerPage: number
+  totalItems: number
+  onPageChange: (page: number) => void
+  onItemsPerPageChange: (items: number) => void
+}
+
+export default function PaginationControls({
+  currentPage,
+  totalPages,
+  itemsPerPage,
+  totalItems,
+  onPageChange,
+  onItemsPerPageChange
+}: PaginationControlsProps) {
+  if (totalItems === 0) return null
+
+  return (
+    <div className="bg-white rounded-[15px] shadow-drop p-4 flex flex-col sm:flex-row items-center justify-between gap-4">
+      <div className="flex items-center gap-3">
+        <label className="text-sm font-medium text-gray-700">Hiển thị:</label>
+        <select
+          value={itemsPerPage}
+          onChange={(e) => onItemsPerPageChange(Number(e.target.value))}
+          className="px-3 py-2 border border-gray-300 rounded-[8px] bg-white text-gray-900 focus:outline-none focus:ring-2 focus:ring-navy-1 focus:border-transparent"
+        >
+          <option value={20}>20</option>
+          <option value={30}>30</option>
+          <option value={50}>50</option>
+        </select>
+        <span className="text-sm text-gray-600">
+          / trang (Tổng: {totalItems} hộ khẩu)
+        </span>
+      </div>
+      
+      <div className="flex items-center gap-2">
+        <button
+          onClick={() => onPageChange(Math.max(1, currentPage - 1))}
+          disabled={currentPage === 1}
+          className="p-2 rounded-[8px] border border-gray-300 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+        >
+          <ChevronLeft className="h-5 w-5" />
+        </button>
+        
+        <div className="flex items-center gap-1">
+          {Array.from({ length: Math.min(5, totalPages) }, (_, i) => {
+            let pageNum: number
+            if (totalPages <= 5) {
+              pageNum = i + 1
+            } else if (currentPage <= 3) {
+              pageNum = i + 1
+            } else if (currentPage >= totalPages - 2) {
+              pageNum = totalPages - 4 + i
+            } else {
+              pageNum = currentPage - 2 + i
+            }
+            
+            return (
+              <button
+                key={pageNum}
+                onClick={() => onPageChange(pageNum)}
+                className={`px-3 py-1 rounded-[8px] text-sm font-medium transition-colors ${
+                  currentPage === pageNum
+                    ? 'bg-navy-1 text-white'
+                    : 'bg-white text-gray-700 border border-gray-300 hover:bg-gray-50'
+                }`}
+              >
+                {pageNum}
+              </button>
+            )
+          })}
+        </div>
+        
+        <button
+          onClick={() => onPageChange(Math.min(totalPages, currentPage + 1))}
+          disabled={currentPage === totalPages}
+          className="p-2 rounded-[8px] border border-gray-300 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+        >
+          <ChevronRight className="h-5 w-5" />
+        </button>
+        
+        <span className="text-sm text-gray-600 ml-2">
+          Trang {currentPage} / {totalPages}
+        </span>
+      </div>
+    </div>
+  )
+}
+
